+++ src/main/java/com/tylerholcombe/leagueit/league/data/Player.java
@@ -10,14 +10,13 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 @Entity
 public class Player {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long playerId;
-    private Integer rating;
+    private Double rating;
     private Integer wins;
     private Integer losses;
     private Integer ties;
@@ -36,11 +35,11 @@ public class Player {
         this.playerId = playerId;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
